refactor: change tiles to responsive
