+++ src/new-client/.env.development
@@ -1,4 +1,4 @@
-REACT_APP_BROKER_HOST=web-dev.adaptivecluster.com
+REACT_APP_BROKER_HOST=web-demo.adaptivecluster.com
 REACT_APP_BROKER_PORT=80
 REACT_APP_SYMPHONY_APP_ID=reactiveTrader-dev
 ENVIRONMENT_NAME=local
