+++ src/new-client/src/App/LiveRates/Tiles/Notional.tsx
@@ -1,10 +1,6 @@
 import React from "react"
-import {
-  InputWrapper,
-  CurrencyPairSymbol,
-  Input,
-  ResetInputValue,
-} from "./styled"
+import { CurrencyPairSymbol, Input, ResetInputValue } from "./styled"
+import { InputWrapper } from "./responsiveWrappers"
 import { formatAsWholeNumber } from "utils/formatNumber"
 interface Props {
   currencyPairBase: string
